[ Add ] : AttributeString Builder
