+++ Sources/WaitMerge/OhluluExtensions.swift
@@ -8,136 +8,6 @@
 
 import UIKit
 
-public struct AttributeStringBuilder {
-    
-    private var attString: NSMutableAttributedString = NSMutableAttributedString()
-    private(set) var baseAttributed: [NSAttributedString.Key: Any] = [:]
-    private(set) var specialAttributed: [NSAttributedString.Key: Any] = [:]
-    
-    public mutating func setBase(attribute: [NSAttributedString.Key: Any]) -> Self {
-        baseAttributed = attribute
-        return self
-    }
-    
-    public mutating func setSepcial(attribute: [NSAttributedString.Key: Any]) -> Self {
-        specialAttributed = attribute
-        return self
-    }
-    
-    public mutating func setBase(
-        text: String,
-        attribute: [NSAttributedString.Key: Any]? = nil)
-        -> Self {
-            attString = NSMutableAttributedString(string: text, attributes: attribute ?? baseAttributed)
-            return self
-    }
-    
-    public func setSpecial(
-        text specialText: String,
-        attribute: [NSAttributedString.Key: Any]? = nil)
-        -> Self {
-            let specialRange = attString.string.ranges(of: specialText)
-                .map { NSRange($0, in: attString.string)}
-            specialRange.forEach {
-                attString.addAttributes(attribute ?? specialAttributed, range: $0)
-            }
-            return self
-    }
-    
-    public func appendSpecial(
-        text: String,
-        attribute: [NSAttributedString.Key: Any]? = nil)
-        -> Self {
-            attString.append(
-                NSMutableAttributedString(string: text, attributes: attribute ?? specialAttributed)
-            )
-            return self
-    }
-    
-    public func appendBase(
-        text: String,
-        attribute: [NSAttributedString.Key: Any]? = nil)
-        -> Self {
-            attString.append(
-                NSMutableAttributedString(string: text, attributes: attribute ?? baseAttributed)
-            )
-            return self
-    }
-    
-    public func build() -> NSMutableAttributedString {
-        return attString
-    }
-}
-
-public extension NSAttributedString {
-    
-    /**
-     ### Usage Exsample ###
-     ```
-     let att = NSAttributedString.builder
-         .setBase(text: *SwiftMininos,
-                  attribute: [
-                     .font: UIFont.systemFont(ofSize: 16, weight: .regular),
-                     .foregroundColor: UIColor.red])
-         .setSpecial(text: "*", attribute: [.foregroundColor: UIColor.black])
-         .build()
-     ```
-     */
-    static var builder: AttributeStringBuilder {
-        get { return AttributeStringBuilder() }
-        set { }
-    }
-}
-
-private extension String {
-    func ranges(of searchString: String) -> [Range<String.Index>] {
-        let _indices = indices(of: searchString)
-        let count = searchString.count
-        return _indices.map({ index(startIndex, offsetBy: $0)..<index(startIndex, offsetBy: $0+count) })
-    }
-}
-
-private extension String {
-    func indices(of occurrence: String) -> [Int] {
-        var indices = [Int]()
-        var position = startIndex
-        while let range = range(of: occurrence, range: position..<endIndex) {
-            let i = distance(from: startIndex,
-                             to: range.lowerBound)
-            indices.append(i)
-            let offset = occurrence.distance(from: occurrence.startIndex,
-                                             to: occurrence.endIndex) - 1
-            guard let after = index(range.lowerBound,
-                                    offsetBy: offset,
-                                    limitedBy: endIndex) else {
-                                        break
-            }
-            position = index(after: after)
-        }
-        return indices
-    }
-}
-
-private extension String {
-    func nsRange(from range: Range<String.Index>) -> NSRange {
-        let from = range.lowerBound.samePosition(in: utf16)!
-        let to = range.upperBound.samePosition(in: utf16)!
-        return NSRange(location: utf16.distance(from: utf16.startIndex, to: from ),
-                       length: utf16.distance(from: from, to: to))
-    }
-}
-
-private extension NSRange {
-    init(_ range: Range<String.Index>, in string: String) {
-        self.init()
-        let startIndex = range.lowerBound.samePosition(in: string.utf16)!
-        let endIndex = range.upperBound.samePosition(in: string.utf16)!
-        self.location = string.distance(from: string.startIndex,
-                                        to: range.lowerBound)
-        self.length = string.distance(from: startIndex, to: endIndex)
-    }
-}
-
 public extension CGSize {
     
     /// create size with square
